pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.2",
    "alembic>=1.17.2",
    "fastapi>=0.122.0",
    "google-api-python-client>=2.187.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.3",
    "langchain>=1.1.0",
    "langchain-core>=1.1.0",
    "langchain-openai>=1.1.0",
    "openai>=2.8.1",
    "pydantic-settings>=2.12.0",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "telegram>=0.0.1",
    "telethon>=1.42.0",
    "uvicorn>=0.38.0",
]
